chore: Restructure Editor sidebar (#3057)
